Fixes: 	1. Changed the http response error throws from util 	2. Added the logic to delete the s3 image after upload the new one

+++ api/controllers/users.py
@@ -230,30 +230,36 @@ def upload_user_profile_picture(credentials, id):
     # validating file extension
     if ext not in ('.png', '.jpg', '.jpeg'):
         bottle.abort(400, 'File extension not allowed.')
-    # validate and read the file
-    img_byte_str = util.read_file_content(upload.file, app.config['profile_img_max_size'])
-
     # validate user detail
     if not util.is_current_user(uid=id, credentials=credentials):
         bottle.abort(403, 'Not authorized to update profile')
+
     try:
+        # validate and read the file
+        img_byte_str = util.read_file_content(upload.file, app.config['profile_img_max_size'])
         user = u.get(id)
         if not user:
             bottle.abort(403, 'Not authorized to update profile picture')
-        pic_url = user.avatar_url
-        # removing old avatar picture from s3 bucket
-        if pic_url and pic_url != '':
-            s3_service.delete_object(Bucket=app.config['aws_s3_bucket_name'], Key='profile/' + pic_url.split('/')[-1])
+
         # upload new avatar picture with new uuid into s3 bucket
         file_name = str(uuid.uuid4()) + ext
-        s3_service.put_object(Body=img_byte_str, Bucket=app.config['aws_s3_bucket_name'], Key='profile/' + file_name,
-                              ACL='public-read', ContentType=upload.content_type)
-        # update avatar s3 ur in user object
-        base_url = app.config['aws_s3_profile_base_url'] + '/profile/' + file_name
-        u.update(user.id, {'avatar_url': base_url})
-        return json.dumps(user.to_dict())
+        response = s3_service.put_object(Body=img_byte_str, Bucket=app.config['aws_s3_bucket_name'],
+                                         Key='profile/' + file_name, ACL='public-read', ContentType=upload.content_type)
+        if response['ResponseMetadata']['HTTPStatusCode'] == 200:
+            # removing old avatar picture from s3 bucket
+            pic_url = user.avatar_url
+            if pic_url and pic_url != '':
+                s3_service.delete_object(Bucket=app.config['aws_s3_bucket_name'], Key='profile/' + pic_url.split('/')[-1])
+            # update avatar s3 ur in user object
+            base_url = app.config['aws_s3_profile_base_url'] + '/profile/' + file_name
+            u.update(user.id, {'avatar_url': base_url})
+            return json.dumps(user.to_dict())
+        else:
+            raise Exception('Avatar S3 upload failed')
     except db.orm.exc.NoResultFound as ex:
         bottle.abort(404, 'User Not found')
+    except AssertionError as ex:
+        bottle.abort(413, str(ex))
     except Exception as ex:
         logging.exception('Could not upload user profile picture: %s' % (ex))
-        bottle.abort(400, 'Could not upload user profile picture')
+        bottle.abort(400, 'Could not upload user profile picture: %s' % (ex))
